Added in deohering and postwhitening
Postwhitening allows for "structure" frequencies to be found (chi-squared test). However, the addition of postwhitening and decohering has lead to improper answers from other functions. Need to return the eigencoefficients and weights with the transfer function.

A complete re-write of this code, at some point, needs to be done taking into account the proper order of operations performed.